+++ velstra-cloud-nodeagent/src/lib.rs
@@ -79,6 +79,7 @@ pub mod guests;
 pub mod host;
 pub mod hostcpu;
 pub mod hostfs;
+pub mod installed;
 pub mod localnet;
 pub mod lvm_pool;
 pub mod metadata;